+++ Changelog.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 1.1.0
+
+- Added fuzzyfyr for CMS Pages
+- Added fuzzyfyr for CMS Blocks
+- Extended fuzzyfyr for Customers with address fuzzing
+
 ## 1.0.1
 
 - Extended documentation